feat ISA-10: IntergrationManager 추가 및 DailyRecordManager에서 토큰 재발급 로직 통합, EmotionView에서 이미지 처리 개선
